README and sample input update

+++ install.sh
@@ -1,4 +1,6 @@
-rm -rf build
+if [ -d build ]; then
+    rm -rf build
+fi
 cmake -B build -DCMAKE_INSTALL_PREFIX=install
 cd build
 make -j 12
